Optimize transfer of video frames to textures.
This does two things:
1) Use pixel buffer objects, if available, and
2) Use bgra instead of rgb24 pixel format for non-yuv420p videos.

The first change allows the OpenGL driver to avoid a data copy, see the
ARB_pixel_buffer_object spec. This leads to a factor 2 speedup with the
proprietary Nvidia drivers in my test with a 2x1080p stereo video.
With the free ATI r600 driver, it currently leads to a factor 2 slowdown
instead, but this is something that will probably be fixed some day.