Improve masking output modes.
The output modes even-odd-rows, even-odd-columns, and checkerboard are now
implemented by the fragment shader. The old method is used as a fallback when
fragment shaders are not available.

This shader implements proper filtering to account for masked pixels.

For even-odd-rows:    [0 1 0]
                      [0 2 0]
		      [0 1 0]

For even-odd-columns: [0 0 0]
                      [1 2 1]
		      [0 0 0]

For checkerboard:     [0 1 0]
                      [1 4 1]
		      [0 1 0]

As a side effect, the masking modes now work again with the free ATI drivers on
Linux.