Work around OpenGL driver bugs again.
This time the workaround is for the NVIDIA 256.* series on Linux.

The GLSL compiler failed on the original version of the code. A variant of the
code with the three local variables sr, sg, sb but an inlined computation
compiled successfully, but resulted in wrong colors, probably due to invalid
optimizations.