Media object: never drop video packets.
This partially reverts commit c5a247c7c3783b28182aef0e63e366d8aeefc574. We
cannot drop video packets because of their interdependencies: it would mess up
decoding.

So we restrict this packet-dropping-because-of-missing-timestamp to audio
packets. Such packets should not have interdependencies, and timestamps are
more important there because the audio clock is the master clock.