Fix sleeping when there is nothing to do.
The original code only slept when there were no events to process, as
determined by video_output::has_events().

However, QApplication::hasPendingEvents() seems to be broken and always returns
true, even if we just called QApplication::processEvents() which is supposed to
process and remove all events.

This means we have no way to determine if there are events or not. Therefore,
this change removes video_output::has_events() completely, and we just always
process events and sleep afterwards.