Initialize subtitle renderer in a separate thread in the background.
If fontconfig needs to build a cache first (mostly on Windows and Mac OS X),
the initialization of libass can take a long time. Therefore we do it in a
background thread and hope that it is finished when the first subtitle has
to be displayed.

If it is not, we wait for it and display a message to the user. Video
playback is paused for the duration of the wait so that no frames are dropped.