Add support for an adjustable audio delay.
The delay is useful if audio and video are not exactly in sync.
This is a workaround for problems described in #35103.

There are some limitations for A/V sync because we do not have precise
presentation time stamps for OpenGL frames, especially in the presence of
compositing window managers. See https://savannah.nongnu.org/bugs/?35103 for
a discussion.