Media object: drop frames with unexpected sizes.
Drop frames that do not have the same dimensions as the initial frame in the
stream.

This fixes crashes with streams that change resolution.

And it also fixes Sony MPO files which store thumbnails inside: the thumbnails
are ignored, and only the full frames are seen.

In the long run, Bino should probably support resolution changes during
playback. When it does, the problem of thumbnails in image files will have to
be revisited.