Big restructuring, part 3.
This changes the video output from request-based rendering to loop-based
rendering in a separate GL thread.

This is necessary for future features such as alternating left/right output.

It also simplifies things because updates do not need to be triggered manually
anymore, e.g. on parameter changes or window movement.