Video output: improve quality==0 path.
Instead of storing linear RGB in a GL_RGB8 texture, which leads to problems in
dark areas, we now store SRGB in GL_RGB8 and accordingly deactivate the
rgb_to_srgb() function in the render shader.

Of course interpolating / filtering SRGB image data instead of linear RGB
introduces an error, but that is the price to pay for quality==0.

This new method is even cheaper than the previous method and introduces smaller
errors.

It was suggested by Frédéric Devernay.