Convert JPEG variants to something digestible before playback
This works around the problem that the Qt Multimedia backends try
hardware acceleration on decoding JPEGs, which fails if the JPEG does
not fit in the limitations of the accelerator. This is regularly the
case for photographs.

Also, this allows manual interpretation of MPO files, which are not reliably
supported by Qt Multimedia backends.