Reorganize meta data detection, and use ffprobe to get stereo-3d and surround meta data.
QMediaMetaData does not provide stereo-3d and surround meta data, and the bug
report about that is being ignored, see
<https://bugreports.qt.io/browse/QTBUG-129458>.

To work around that, Bino now runs the ffprobe tool (if available) to gather
all meta data in a QJsonDocument so that it can interpret stereo-3d and
surround meta data.

If ffprobe is not available or fails for some reason, Bino falls back to
guessing, as before.

The reason to run ffprobe instead of linking with libavformat and
programmatically extracting the information is that the FFmpeg API breaks
constantly and keeping up takes too much time (this is the experience from the
bino-1.x series). Furthermore, building with libavformat complicates the build
process considerably and requires depending on specific versions.